requirements.txt
Django==5.2.8
langgraph==1.0.3
langchain==1.0.8
langchain-openai==1.0.3
pandas==2.3.3
python-dotenv==1.2.1
openai==2.8.1
python-gemini-api==2.4.12
claude-api-py==0.0.5
streamlit==1.51.0
